Add theme stubs and logging
It seems to be rather hard to query the firefox colour scheme -
ideally we'd just figure out the colours, and update the config page
and new tab page to fit in. Unfortunately at least without a theme
installed there doesn't seem to be any output - but even if there were
it's questionable how usueful the result would be: According to the
documentation information is about a theme applied to the current tab,
after it has been applied.